requirements.txt
# College Event Statistics Portal — Python dependencies
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
